adding and updating tuto screenshot
