+++ README.md
@@ -8,9 +8,10 @@ The various tutorials are:
 3. [tutorial 3](./tutorial_3/README.md) How to use HPP in manufacturing.
 4. [tutorial 4](./tutorial_4/README.md) How to control the trajectory of a tool.
 5. [tutorial 5](./tutorial_5/README.md) How to optimize and time-parameterize paths.
-6. [tutorial 6](./tutorial_6/README.md) How to execute motions on a simulated robot.
-7. [tutorial 7](./tutorial_7/README.md) Pick and place with gripper, introducing pre and post-actions.
-8. [tutorial 8](./tutorial_8/README.md) A more efficient way to assign pre or post-actions.
+6. [tutorial 6](./tutorial_6/README.md) How to use the rviz2 vizualization.
+7. [tutorial 7](./tutorial_7/README.md) How to execute motions on a simulated robot.
+8. [tutorial 8](./tutorial_8/README.md) Pick and place with gripper, introducing pre and post-actions.
+9. [tutorial 9](./tutorial_9/README.md) A more efficient way to assign pre or post-actions.
 
 [![Pipeline status](https://gitlab.laas.fr/humanoid-path-planner/hpp_tutorial/badges/master/pipeline.svg)](https://gitlab.laas.fr/humanoid-path-planner/hpp_tutorial/commits/master)
 [![Coverage report](https://gitlab.laas.fr/humanoid-path-planner/hpp_tutorial/badges/master/coverage.svg?job=doc-coverage)](https://gepettoweb.laas.fr/doc/humanoid-path-planner/hpp_tutorial/master/coverage/)
