+++ tutorial_6/README.md
@@ -1,8 +1,8 @@
-# How to have a rviz2 visualization
+# How to use a rviz2 visualization
 
 ## Prerequisite
 
-Having completed [tutorial 5](../tutorial_5/README.md).
+Having completed [tutorial 2](../tutorial_2/README.md).
 
 ## Overview
 
@@ -52,7 +52,7 @@ export the package path
 export ROS_PACKAGE_PATH=/home/user/devel/src/:$ROS_PACKAGE_PATH
 ```
 
-## Initializing the problem
+## Initializing the viewer
 
 In the docker container, cd into `tutorial_6` directory and run:
 
@@ -60,8 +60,14 @@ In the docker container, cd into `tutorial_6` directory and run:
 python -i init.py
 ```
 
-The script is identical to tutorial 5 up to the computation of paths `p1`, `p2`, `fullPath`.
-The only difference is that it uses `RVizVisualizer` instead of `Viewer`.
+The script is identical to tutorial 2 up to the computation of path `p`
+The only difference is that it uses `RVizVisualizer` instead of `viser Viewer`.
+
+```python
+v = Viewer()
+v.initViewer(robot=robot)
+
+```
 
 ## Configuring RViz2
 
@@ -77,8 +83,8 @@ In RViz2, set the **Fixed Frame** to `world` (in the *Global Options* panel).
 Set the Frame Rate to 300
 
 For each model loaded with `urdf.loadModel`, add a **RobotModel** display and set its
-`Description Topic` to the topic published for that model (e.g. `/staubli/robot_description`,
-`/plate/robot_description`, `/obstacle/robot_description`). The prefix matches the name
+`Description Topic` to the topic published for that model (e.g. `/panda/robot_description`,
+`/ground/robot_description`, `/box/robot_description`). The prefix matches the name
 passed to `urdf.loadModel`.
 
 Add a **TF** display to visualize all frame transforms published by the viewer.
@@ -87,16 +93,28 @@ Disable TF arrows and put Frame Time Out to 1e+07
 
 There is a rviz config file on hpp_tutorial/launch/tuto6.rviz
 ```bash
-rviz2 -d hpp_tutorial/launch/tuto6.rviz
+rviz2 -d hpp_tutorial/launch/config.rviz
 ```
 
 Call `v(q_init)` in the Python terminal to place all objects in their initial configuration.
 The scene should now appear in RViz2.
 
+![Control of the view](figures/q_init_rviz2.png)
+
+## Visualizing a configuration
+
+```python
+v(q_init)
+```
+Use v(config) to display a configuration on rviz2 like is done with viser. 
+
 ## Visualizing a path
 
 Add the **Trajectory** plugin (from the HPP RViz2 plugins) to RViz2. A control panel
 will appear at the bottom of the screen.
+
+![Control of the view](figures/trajectory_display_select.png)
+
 
 Load a path from the Python terminal:
 
@@ -109,26 +127,29 @@ The panel lets you slide through the path parameter to inspect any intermediate 
 To display the spatial trace of a frame along the path, use:
 
 ```python
-v.displayPath(p1, target_frame="staubli/tooltip")
+v.displayPath(p1, target_frame="panda/gripper")
 ```
 
 You can also trigger this from RViz2 by entering the frame name directly in the
 DisplayTrajectory panel.
 
+![Control of the view](figures/trajectory_panel_up.png)
+
+
 ## Adding waypoints
 
 Add the **Waypoint** tool from the RViz2 toolbar (installed with the HPP RViz2 plugins).
-Add the **DisplayWaypoint** display to visualize the waypoints in the scene.
+Add the **Waypoint** display to visualize the waypoints in the scene.
 
 Waypoints can be placed in three ways:
 
 - **Interactively**: select the Waypoint tool in the toolbar, then click in the 3D view.
   Drag the waypoint with the interact tool for moving it or
-  Right-click a waypoint marker and choose *Edit position* to adjust it numerically with the intrect tool too
+  Right-click a waypoint marker and choose *Edit position* to adjust it numerically with the interact tool too
 
 - **From a named frame** (Python):
   ```python
-  v.addWaypointFromFrame("staubli/tooltip")
+  v.addWaypointFromFrame("panda/gripper")
   ```
   This publishes the current pose of the given frame as a waypoint.
 
@@ -137,6 +158,7 @@ Waypoints can be placed in three ways:
   v.addWaypoint(xyz=[0.8, 0.0, 1.0], quat_xyzw=[0.0, 0.0, 0.0, 1.0])
   ```
 
+![Alt Text](figures/waypoint.gif)
 ## Summary of viewer methods
 
 | Method | Description |
